Adding users logic

+++ src/hooks/use-process-api-query.tsx
@@ -8,6 +8,7 @@ import { toast } from 'sonner'
 import { getAvatarRandom } from '../utils/avatar-util'
 import { AccountInfo } from '@azure/msal-browser'
 import { ProcessBase } from '../pages/process/interfaces/process-base'
+import { useUserRole } from './use-user-role'
 
 interface useProcessApiQueryProps {
   processId?: number
@@ -42,6 +43,7 @@ export const useProcessApiQuery = ({
   enabled = true,
 }: useProcessApiQueryProps) => {
   const [processes, setProcesses] = useState<Process[]>([])
+  const { isAdmin, isUser, isObserver } = useUserRole()
   const { accounts } = useMsal()
   const account = accounts[0]
 
@@ -49,8 +51,11 @@ export const useProcessApiQuery = ({
     useProcessApi()
   const { data, isLoading, error } = useQuery({
     queryKey: ['processes', processId],
-    queryFn: () =>
-      getProceses({
+    queryFn: () => {
+      if (!isAdmin() && !isUser() && !isObserver())
+        throw new Error('You are not authorized to get processes')
+
+      return getProceses({
         processId,
         includeCategories,
         includeWorkItems,
@@ -58,17 +63,23 @@ export const useProcessApiQuery = ({
         includeTags,
         includeProcessUsers,
         omitWorkItemsAbandoned,
-      }),
+      })
+    },
     staleTime: 300000,
     cacheTime: 600000,
     enabled: enabled,
-    onError: () => {
-      toast.error('An error occurred while trying to get processes')
+    onError: (error: Error) => {
+      toast.error(
+        error?.message || 'An error occurred while trying to get processes'
+      )
     },
   })
 
   const mutationAddProcess = useMutation({
     mutationFn: (process: Process) => {
+      if (!isAdmin())
+        throw new Error('You are not authorized to create a process')
+
       return createProcess({
         name: process.name,
         description: process.description,
@@ -87,15 +98,18 @@ export const useProcessApiQuery = ({
       ])
       toast.success('Process created successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to create a process, please try again'
+        error.message ||
+          'An error occurred while trying to create a process, please try again'
       )
     },
   })
 
   const mutationEditProcess = useMutation({
     mutationFn: (process: Process) => {
+      if (!isAdmin())
+        throw new Error('You are not authorized to update a process')
       return updateProcess({
         id: process.id,
         name: process.name,
@@ -113,24 +127,28 @@ export const useProcessApiQuery = ({
       )
       toast.success('Process updated successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to update a process, please try again'
+        error.message ||
+          'An error occurred while trying to update a process, please try again'
       )
     },
   })
 
   const mutationDeleteProcess = useMutation({
     mutationFn: (id: number) => {
+      if (!isAdmin())
+        throw new Error('You are not authorized to delete a process')
       return deleteProcess(id)
     },
     onSuccess: (data: ProcessBase) => {
       setProcesses(processes.filter((process) => process.id !== data.id))
       toast.success('Process deleted successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to delete a process, please try again'
+        error.message ||
+          'An error occurred while trying to delete a process, please try again'
       )
     },
   })
