+++ src/hooks/use-process-user-api-query.tsx
@@ -10,6 +10,7 @@ import { getAvatarRandom } from '../utils/avatar-util'
 import { ProcessUserBase } from '../pages/sherpas-process/interfaces/process-user-base'
 import { Process } from '../pages/process/interfaces/process'
 import { ProcessUserForm } from '../pages/sherpas-process/interfaces/process-user-form'
+import { useUserRole } from './use-user-role'
 
 interface useProcessUserApiQueryProps {
   proccesses: Process[]
@@ -73,6 +74,7 @@ export const useProcessUserApiQuery = ({
   enabled = true,
 }: useProcessUserApiQueryProps) => {
   const [processesUsers, setProcessesUsers] = useState<ProcessUser[]>([])
+  const { isAdmin, isUser, isObserver } = useUserRole()
   const { accounts } = useMsal()
   const account = accounts[0]
   const {
@@ -84,18 +86,31 @@ export const useProcessUserApiQuery = ({
 
   const { data, isLoading, error } = useQuery({
     queryKey: ['processesUsers', processId],
-    queryFn: () =>
-      getProcessUsers({ processId, userId, includeProcess, includeUsers }),
+    queryFn: () => {
+      if (!isAdmin() && !isUser() && !isObserver())
+        throw new Error('You are not authorized to get processes users')
+      return getProcessUsers({
+        processId,
+        userId,
+        includeProcess,
+        includeUsers,
+      })
+    },
     staleTime: 300000,
     cacheTime: 600000,
     enabled: enabled,
-    onError: () => {
-      toast.error('An error occurred while trying to get processes users')
+    onError: (error: Error) => {
+      toast.error(
+        error?.message ||
+          'An error occurred while trying to get processes users'
+      )
     },
   })
 
   const mutationAddProcessUser = useMutation({
     mutationFn: (processUser: ProcessUserForm) => {
+      if (!isAdmin() && !isUser())
+        throw new Error('You are not authorized to create a process user')
       return createProcessUser({
         processId: processUser.processId,
         user: processUser.user ?? createBaseUser(account),
@@ -116,15 +131,18 @@ export const useProcessUserApiQuery = ({
       ])
       toast.success('Process user created successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to create a process user, please try again'
+        error?.message ||
+          'An error occurred while trying to create a process user, please try again'
       )
     },
   })
 
   const mutationEditProcessUser = useMutation({
     mutationFn: (processUser: ProcessUser) => {
+      if (!isAdmin() && !isUser())
+        throw new Error('You are not authorized to update a process user')
       return updateProcessUser({
         processId: processUser.processId,
         user: processUser.user ?? createBaseUser(account),
@@ -153,15 +171,18 @@ export const useProcessUserApiQuery = ({
       )
       toast.success('Process user updated successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to update a process user, please try again'
+        error?.message ||
+          'An error occurred while trying to update a process user, please try again'
       )
     },
   })
 
   const mutationDeleteProcessUser = useMutation({
     mutationFn: (processUser: ProcessUser) => {
+      if (!isAdmin() && !isUser())
+        throw new Error('You are not authorized to delete a process user')
       return deleteProcessUser({
         processId: processUser.processId,
         userId: processUser.user?.id || processUser.userId,
@@ -177,9 +198,10 @@ export const useProcessUserApiQuery = ({
       )
       toast.success('Process user deleted successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to delete a process user, please try again'
+        error?.message ||
+          'An error occurred while trying to delete a process user, please try again'
       )
     },
   })
