+++ src/hooks/use-tag-api-query.tsx
@@ -3,24 +3,33 @@ import { Tag } from '../pages/tags/interfaces/tag'
 import { useMutation, useQuery } from 'react-query'
 import { useTagsApi } from './use-tags-api'
 import { toast } from 'sonner'
+import { useUserRole } from './use-user-role'
 
 export const useTagsApiQuery = (enabled: boolean = true) => {
   const [tags, setTags] = useState<Tag[]>([])
 
   const { getTags, createTag, updateTag, deleteTag } = useTagsApi()
+  const { isAdmin, isUser, isObserver } = useUserRole()
   const { data, isLoading, error } = useQuery({
     queryKey: 'tags',
-    queryFn: () => getTags(),
+    queryFn: () => {
+      if (!isAdmin() && !isUser() && !isObserver())
+        throw new Error('You are not authorized to get tags')
+      return getTags()
+    },
     staleTime: 300000,
     cacheTime: 600000,
-    onError: () => {
-      toast.error('An error occurred while trying to get tags')
+    onError: (error: Error) => {
+      toast.error(
+        error?.message || 'An error occurred while trying to get tags'
+      )
     },
     enabled: enabled,
   })
 
   const mutationAddTag = useMutation({
     mutationFn: (tag: Tag) => {
+      if (!isAdmin()) throw new Error('You are not authorized to create tags')
       return createTag({
         name: tag.name,
         description: tag.description,
@@ -31,15 +40,17 @@ export const useTagsApiQuery = (enabled: boolean = true) => {
       setTags([...tags, data])
       toast.success('Tag created successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to create a tag, please try again'
+        error?.message ||
+          'An error occurred while trying to create a tag, please try again'
       )
     },
   })
 
   const mutationEditProcess = useMutation({
     mutationFn: (tag: Tag) => {
+      if (!isAdmin()) throw new Error('You are not authorized to update tags')
       return updateTag({
         id: tag.id,
         name: tag.name,
@@ -58,24 +69,27 @@ export const useTagsApiQuery = (enabled: boolean = true) => {
       )
       toast.success('Tag updated successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to update a tag, please try again'
+        error?.message ||
+          'An error occurred while trying to update a tag, please try again'
       )
     },
   })
 
   const mutationDeleteProcess = useMutation({
     mutationFn: (id: number) => {
+      if (!isAdmin()) throw new Error('You are not authorized to delete tags')
       return deleteTag(id)
     },
     onSuccess: (data) => {
       setTags(tags.filter((tag) => tag.id !== data.id))
       toast.success('Tag deleted successfully')
     },
-    onError: () => {
+    onError: (error: Error) => {
       toast.error(
-        'An error occurred while trying to delete a tag, please try again'
+        error?.message ||
+          'An error occurred while trying to delete a tag, please try again'
       )
     },
   })
